backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
opencv-python
torch
torchvision
numpy
requests
Pillow
transformers>=4.30.0
supervision>=0.16.0
PyYAML
addict
yapf
timm
open_clip_torch
anomalib
pytorch-lightning
openai
scikit-learn>=1.0.0
joblib>=1.0.0


# Google Cloud Storage Dependencies
google-cloud-storage>=2.10.0
google-auth>=2.17.0

# Additional cloud dependencies
google-cloud-core>=2.3.0

# Environment and configuration
python-dotenv>=1.0.0

# Production server (for Cloud Run deployment)
gunicorn>=21.0.0
uvicorn[standard]>=0.24.0